Add smoke tests for sbt-tapiro
