+++ tapiro/sbt-tapiro/src/sbt-test/sbt-tapiro/simple/project/build.properties
@@ -1,0 +1,1 @@
+sbt.version=1.3.8
