+++ tapiro/sbt-tapiro/src/sbt-test/sbt-tapiro/simple/project/plugins.sbt
@@ -1,3 +1,4 @@
+addSbtPlugin("io.buildo" %% "sbt-buildo" % "0.11.5")
 sys.props.get("plugin.version") match {
   case Some(v) => addSbtPlugin("io.buildo" % "sbt-tapiro" % v)
   case _       => sys.error("""|The system property 'plugin.version' is not defined.
