+++ tapiro/sbt-tapiro/src/sbt-test/sbt-tapiro/simple/src/main/resources/logback.xml
@@ -1,0 +1,3 @@
+<configuration>
+  <root level="info" />
+</configuration>
