--- tapiro/sbt-tapiro/src/sbt-test/sbt-tapiro/simple/src/main/scala/Api.scala
@@ -1,20 +1,0 @@
-package testo
-
-import scala.annotation.StaticAnnotation
-class query extends StaticAnnotation
-class command extends StaticAnnotation
-
-case class CustomError(message: String)
-
-trait Controller[F[_]] {
-  @query
-  def queryExample(
-    intParam: Int,
-    stringParam: String,
-  ): F[Either[CustomError, CustomModel]]
-
-  @command
-  def commandExample(
-    body: CustomModel,
-  ): F[Either[String, String]]
-}
