+++ tapiro/sbt-tapiro/src/sbt-test/sbt-tapiro/simple/src/main/scala/CustomModel.scala
